app/requirements.txt
dash
Pillow
numpy
plotly
